requirements.txt
python-dotenv==1.0.0
Requests==2.31.0
st_pages==0.4.1
streamlit==1.22.0
streamlit_extras==0.2.7
